backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
pydantic==2.10.3
supabase==2.10.0
python-multipart==0.0.19
python-dotenv==1.0.1
